Update Homepage, and About me Page
Redoing to homepage so that it is more image than text based. And with the about me, I fused the about me and resume pages into on and Used a design style rel of the about me. Next up adding the resume link below the image and than the indivisual project pages.